Initial submit of nanosort
This is a work in progress and contains nanosort code along with a
benchmark vs pdqsort and branchless sort by Gerben Stavenga.

nanosort's small sort is using Gerben's BubbleSort2 for now, this needs
further analysis.

MSVC generates rather suboptimal code for the branchless partition in
nanosort - this will need a compiler fix and/or workarounds in the code.