Replace small sort with a variant of bubble-2 sort
This is semantically similar to the previous implementation but uses a
more canonical 3-element sort network to order elements during runs.

Seems to run at the same speed with clang, but this can be more
efficient when copies are expensive, and results in more uniform codegen
demands wrt the rest of the code.